Create 100 Days challenge
🚀 100 Days of Coding Challenge 🚀

Welcome to the 100 Days of Coding Challenge, your first step toward becoming a confident developer from scratch.

🎯 Challenge Objective

The aim is straightforward: spend part of your day on learning and coding for the next 100 days. By the end, you’ll be well-versed in JavaScript/TypeScript and its uses, turning you into a skilled developer. 🌈